initial commit for the mysite project

+++ yellowpages/spiders/tester.py
@@ -7,22 +7,23 @@ import requests
 
 class neworleansSpider(scrapy.Spider):
     name = "tester"
-    allowed_domains = ['https://www.yellowpages.com']
-
-    def start_requests(self):
-        yield scrapy.Request('https://www.yellowpages.com/search?search_terms=restaurants&geo_location_terms=New%20Orleans%2C%20LA', self.parse)
-        yield scrapy.Request('https://www.yellowpages.com/search?search_terms=restaurants&geo_location_terms=New%20Orleans%2C%20LA&page=2', self.parse)
-        yield scrapy.Request('https://www.yellowpages.com/search?search_terms=restaurants&geo_location_terms=New%20Orleans%2C%20LA&page=3', self.parse)
+    start_urls = ['https://www.yellowpages.com/search?search_terms=restaurants&geo_location_terms=New%20Orleans%2C%20LA',
+    'https://www.yellowpages.com/search?search_terms=restaurants&geo_location_terms=New%20Orleans%2C%20LA&s=default&page=2']
+    #start = 0
+    #rest_ids = []
 
     def parse(self, response):
-        file_name = 'nolaoutputfile38.csv'
+        file_name = 'nolaoutputfile.csv'
         #restaurant_info = response.css('div#main-content')
         restaurants = response.css('div.result')
 
-        out_file = open(file_name, 'w')
-        csv_writer = csv.writer(out_file)
-        csv_writer.writerow( ['name', 'phone_number', 'address', 'city_zip'])
-            #csv_writer.writerow( ['name'] )
+        if os.path.exists(file_name):
+            out_file = open(file_name, 'a')
+            csv_writer = csv.writer(out_file)
+        else:
+            out_file = open(file_name, 'w')
+            csv_writer = csv.writer(out_file)
+            csv_writer.writerow( ['name', 'phone_number', 'address', 'city_zip'])
 
         for restaurant in restaurants:
             name = restaurant.css('span::text').get()
